修改多核实验部分

+++ GraduationProject/Multicore/stream_exp3/USM/Add-USM.cpp
@@ -112,6 +112,7 @@ public:
 int main(int argc, char** argv) {
   BenchmarkApp app(argc, argv);
 
-  app.run<MicroBenchAdd<float, 1>>();
+  //app.run<MicroBenchAdd<float, 1>>();
+  app.run<MicroBenchAdd<double, 1>>();
   return 0;
 }
